requirements.txt
asgiref==3.2.10
braintree
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==2.9.2
dateutils==0.6.11
defusedxml==0.6.0
dj-database-url
Django
django-admin-honeypot
django-allauth
django-autoslug
django-countries
django-crispy-forms==1.9.2
django-honeypot
django-payments
django-randomslugfield
django-ranged-response
django-simple-captcha
django-storages
docutils
gunicorn
idna==2.10
jmespath
oauthlib==3.1.0
Pillow==8.0.1
psycopg2
pycparser==2.20
PyJWT
python-dateutil==2.8.1
python-dotenv==0.15.0
python-utils
python3-openid==3.2.0
pytz==2020.1
requests
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.6.3
South
sqlparse==0.3.1
stripe==2.55.1
suds-jurko==0.6
urllib3==1.25.9
whitenoise
xmltodict==0.12.0
